+++ Tarea2-Grupo23/src/controllers/Personaje_habita_reinoController.js
@@ -4,7 +4,8 @@ import prisma from '../prismaClient.js'
 const getPersonaje_habita_reino = async(req, res) => {
     try {
         const personajes_habitan_reinos = await prisma.personaje_habita_reino.findMany()
-        res.json(personajes_habitan_reinos)
+        if (personajes_habitan_reinos.length === 0){res.status(204).json()}
+        else {res.json(personajes_habitan_reinos)}
     }
     catch (error){res.status(500).json({message: "Internal Server Error"})}
 }
@@ -21,7 +22,8 @@ const getPersonaje_habita_reinoById = async (req, res) => {
                 }
             }
         })
-        res.status(201).json(personaje_habita_reino)
+        if (personaje_habita_reino === null){res.status(404).json({message: "No existe el elemento buscado"})}
+        else {res.json(personaje_habita_reino)}
     }
     catch (error){res.status(500).json({message: "Internal Server Error"})}
 }
@@ -51,19 +53,30 @@ const actualizarPersonaje_habita_reino = async (req, res) => {
     try {
         const {id_personaje, id_reino} = req.params
         const {fecha_registro, es_gobernante} = req.body
-        const personaje_habita_reino = await prisma.personaje_habita_reino.update ({
+        var personaje_habita_reino = await prisma.personaje_habita_reino.findUnique({
             where: {
-                id_personaje_id_reino: {
+                id_personaje_id_reino:{
                     id_personaje: Number(id_personaje),
-                    id_reino: Number(id_reino)
+                    id_reino: Number(id_reino),
                 }
-            },
-            data: {
-                fecha_registro,
-                es_gobernante
             }
         })
-        res.json(personaje_habita_reino)
+        if (personaje_habita_reino === null){res.status(404).json({message: "No existe el elemento buscado"})}
+        else {
+            personaje_habita_reino = await prisma.personaje_habita_reino.update({
+                where: {
+                    id_personaje_id_reino: {
+                        id_personaje: Number(id_personaje),
+                        id_reino: Number(id_reino)
+                    }
+                },
+                data: {
+                    fecha_registro,
+                    es_gobernante
+                }
+            })
+            res.json(personaje_habita_reino)
+        }
     }
     catch (error){res.status(500).json({message: "Internal Server Error"})}
 }
@@ -72,15 +85,26 @@ const actualizarPersonaje_habita_reino = async (req, res) => {
 const eliminarPersonaje_habita_reino = async (req, res) => {
     try {
         const {id_personaje, id_reino} = req.params
-        const personaje_habita_reino = await prisma.personaje_habita_reino.delete({
+        var personaje_habita_reino = await prisma.personaje_habita_reino.findUnique({
             where: {
-                id_personaje_id_reino: {
+                id_personaje_id_reino:{
                     id_personaje: Number(id_personaje),
-                    id_reino: Number(id_reino)
+                    id_reino: Number(id_reino),
                 }
             }
         })
-        res.json({message: "Eliminado con éxito"})
+        if (personaje_habita_reino === null){res.status(404).json({message: "No existe el elemento buscado"})}
+        else {
+            const personaje_habita_reino = await prisma.personaje_habita_reino.delete({
+                where: {
+                    id_personaje_id_reino: {
+                        id_personaje: Number(id_personaje),
+                        id_reino: Number(id_reino)
+                    }
+                }
+            })
+            res.json({message: "Eliminado con éxito"})
+        }
     }
     catch (error){res.status(500).json({message: "Internal Server Error"})}
 }
