+++ Tarea2-Grupo23/src/controllers/ReinosController.js
@@ -3,8 +3,11 @@ import prisma from '../prismaClient.js'
 //GET para obtener todos los elementos de la tabla
 const getReinos = async(req, res) => {
     try {
-        const reinos = await prisma.reinos.findMany()
-        res.json(reinos)
+        const reinos = await prisma.reinos.findMany({
+            include: {defensas: true}
+        })
+        if (reinos.length === 0){res.status(204).json()}
+        else {res.json(reinos)}
     }
     catch (error){res.status(500).json({message: "Internal Server Error"})}
 }
@@ -14,9 +17,11 @@ const getReinoById = async (req, res) => {
     try {
         const {id} = req.params
         const reino = await prisma.reinos.findUnique({
-            where: {id: Number(id)}
+            where: {id: Number(id)},
+            include: {defensas: true}
         })
-        res.json(reino)
+        if (reino === null){res.status(404).json({message: "No existe el elemento buscado"})}
+        else {res.json(reino)}
     }
     catch (error){res.status(500).json({message: "Internal Server Error"})}
 }
@@ -49,20 +54,27 @@ const actualizarReino = async (req, res) => {
     try {
         const {id} = req.params
         const {nombre, ubicacion, superficie, defensas_crear, defensas_conectar, defensas_desconectar} = req.body
-        const reino = await prisma.reinos.update ({
+        var reino = await prisma.reinos.findUnique({
             where: {id: Number(id)},
-            data: {
-                nombre,
-                ubicacion,
-                superficie,
-                defensas: {
-                    create: defensas_crear,
-                    connect: defensas_conectar,
-                    disconnect: defensas_desconectar
-                }
-            }
+            include: {defensas: true}
         })
-        res.json(reino)
+        if (reino === null){res.status(404).json({message: "No existe el elemento buscado"})}
+        else {
+            reino = await prisma.reinos.update ({
+                where: {id: Number(id)},
+                data: {
+                    nombre,
+                    ubicacion,
+                    superficie,
+                    defensas: {
+                        create: defensas_crear,
+                        connect: defensas_conectar,
+                        disconnect: defensas_desconectar
+                    }
+                }
+            })
+            res.json(reino)
+        }
     }
     catch (error){res.status(500).json({message: "Internal Server Error"})}
 }
@@ -71,26 +83,23 @@ const actualizarReino = async (req, res) => {
 const eliminarReino = async (req, res) => {
     try {
         const {id} = req.params
-        //ELIMINACIÓN EN CASCADA: TABLA PERSONAJE_HABITA_REINO
-        const personaje_habita_reino = await prisma.personaje_habita_reino.findMany({
-            where: {id_reino: Number(id)}
+        var reino = await prisma.reinos.findUnique({
+            where: {id: Number(id)},
+            include: {defensas: true}
         })
-        if (personaje_habita_reino.length !== 0){
-            personaje_habita_reino = await prisma.personaje_habita_reino.deleteMany({
+        if (reino === null){res.status(404).json({message: "No existe el elemento buscado"})}
+        else {
+            //ELIMINACIÓN EN CASCADA: TABLA PERSONAJE_HABITA_REINO
+            var personaje_habita_reino = await prisma.personaje_habita_reino.findMany({
                 where: {id_reino: Number(id)}
             })
-        }
-        //ELIMINACIÓN EN CASCADA: TABLA DIPLOMACIAS
-        const diplomacias = await prisma.diplomacias.findMany({
-            where: {
-                OR: [
-                    {id_reino_1: Number(id)},
-                    {id_reino_2: Number(id)}
-                ]
+            if (personaje_habita_reino.length !== 0){
+                personaje_habita_reino = await prisma.personaje_habita_reino.deleteMany({
+                    where: {id_reino: Number(id)}
+                })
             }
-        })
-        if (diplomacias.length !== 0){
-            diplomacias = await prisma.diplomacias.deleteMany({
+            //ELIMINACIÓN EN CASCADA: TABLA DIPLOMACIAS
+            var diplomacias = await prisma.diplomacias.findMany({
                 where: {
                     OR: [
                         {id_reino_1: Number(id)},
@@ -98,12 +107,22 @@ const eliminarReino = async (req, res) => {
                     ]
                 }
             })
+            if (diplomacias.length !== 0){
+                diplomacias = await prisma.diplomacias.deleteMany({
+                    where: {
+                        OR: [
+                            {id_reino_1: Number(id)},
+                            {id_reino_2: Number(id)}
+                        ]
+                    }
+                })
+            }
+            //ELIMINACIÓN
+            reino = await prisma.reinos.delete({
+                where: {id: Number(id)}
+            })
+            res.json({message: "Eliminado con éxito"})
         }
-        //ELIMINACIÓN
-        const reino = await prisma.reinos.delete({
-            where: {id: Number(id)}
-        })
-        res.json({message: "Eliminado con éxito"}) 
     }
     catch (error){
         console.log(error)
